requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
environs==14.1.1
exceptiongroup==1.2.2
fast-depends==2.4.12
fastapi==0.115.8
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.10
frozenlist==1.6.0
greenlet==3.2.1
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
izulu==0.50.0
Jinja2==3.1.6
jsonformatter==0.3.4
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
multidict==6.4.3
mypy-extensions==1.0.0
nats-py==2.10.0
ordered-set==4.1.0
ormsgpack==1.8.0
packaging==24.2
pamqp==3.3.0
propcache==0.3.1
pycron==3.1.2
pydantic==2.11.3
pydantic-settings==2.7.1
pydantic_core==2.33.1
python-dotenv==1.1.0
pytz==2025.1
ruff==0.11.7
segno==1.6.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.45.3
taskiq==0.11.17
taskiq-aiogram==0.3.1
taskiq-dependencies==1.5.6
taskiq-nats==0.5.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.0
validators==0.34.0
watchdog==2.3.1
yarl==1.20.0
zipp==3.21.0
